Added AE6 nemesis mats and dummys for the cards. Dummy cards have no info except for their IDs. Corrected the AE6 abbreviations LOG, LOGA and LOGG to L2, L2A and L2G. Added Maelstrom, including dummy nemesis cards.
Hitpoints are now in string format because of Mother of Rust and Paradox of Myth and Bone. "battle" is now str as well because of Maelstrom, Brama and Z'hana.

Removed an "Â" artifact from the Knight of Shackles' text.

Altered the python code to support Ic cards (no more "upper()") for Maelstrom and L2 nemesis cards. Also there are a few more comments and some additional line-breaks for easier debugging.